requirements.txt
pymysql
flask
faiss
scipy
pillow==4.2.1
grpcio
protobuf
scikit-learn==0.21.3
opencv-contrib-python==3.3.0.10
numpy
matplotlib
torch==1.2.0
requests
joblib
flask-cors
torchvision==0.4.0
efficientnet_pytorch